Refactoring to support multiple databases
