+++ .env
@@ -1,4 +1,4 @@
 # Database configuration
 DB_FOLDER='/database'
-DB_NAME='sqlite.db'
-DB_PATH=${DB_FOLDER}/${DB_NAME}
+STEAM_DB_NAME='sqlite.db'
+STEAM_DB_PATH=${DB_FOLDER}/${STEAM_DB_NAME}
