+++ .github/workflows/docker-publish.yml
@@ -11,8 +11,7 @@ env:
   IMAGE_NAME_PROXY: "${{ github.repository_owner }}/ssgg/proxy"
   IMAGE_NAME_FRONTEND: "${{ github.repository_owner }}/ssgg/frontend"
   IMAGE_NAME_BACKEND: "${{ github.repository_owner }}/ssgg/backend"
-  IMAGE_NAME_DB_HANDLER: "${{ github.repository_owner }}/ssgg/db-handler"
-  IMAGE_NAME_DB_UPGRADER: "${{ github.repository_owner }}/ssgg/db-upgrader"
+  IMAGE_NAME_DB_HANDLER: "${{ github.repository_owner }}/ssgg/steam-db-handler"
 
 
 jobs:
@@ -80,45 +79,29 @@ jobs:
           tags: ${{ steps.meta_backend.outputs.tags }}
           labels: ${{ steps.meta_backend.outputs.labels }}
 
-      # DB Handler image
-      - name: Extract SSGG DB Handler Docker metadata
-        id: meta_db_handler
+      # Steam DB Handler image
+      - name: Extract SSGG Steam DB Handler Docker metadata
+        id: meta_steam_db_handler
         uses: docker/metadata-action@98669ae865ea3cffbcbaa878cf57c20bbf1c6c38
         with:
           images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME_DB_HANDLER }}
 
-      - name: Build and push SSGG DB Handler image
+      - name: Build and push SSGG Steam DB Handler image
         uses: docker/build-push-action@ad44023a93711e3deb337508980b4b5e9bcdc5dc
         with:
           context: backend
-          file: backend/database_handler/Dockerfile
+          file: backend/steam_db_handler/Dockerfile
           push: ${{ github.event_name != 'pull_request' }}
-          tags: ${{ steps.meta_db_handler.outputs.tags }}
-          labels: ${{ steps.meta_db_handler.outputs.labels }}
+          tags: ${{ steps.meta_steam_db_handler.outputs.tags }}
+          labels: ${{ steps.meta_steam_db_handler.outputs.labels }}
 
-      # DB Upgrader image
-      - name: Extract SSGG Frontend Docker metadata
-        id: meta_db_upgrader
-        uses: docker/metadata-action@98669ae865ea3cffbcbaa878cf57c20bbf1c6c38
-        with:
-          images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME_DB_UPGRADER }}
-
-      - name: Build and push SSGG Frontend image
-        uses: docker/build-push-action@ad44023a93711e3deb337508980b4b5e9bcdc5dc
-        with:
-          context: backend
-          file: backend/database_migration/Dockerfile
-          push: ${{ github.event_name != 'pull_request' }}
-          tags: ${{ steps.meta_db_upgrader.outputs.tags }}
-          labels: ${{ steps.meta_db_upgrader.outputs.labels }}
-  
   deploy:
     runs-on: ubuntu-latest
     needs: build
     steps:
       - name: Checkout repository
         uses: actions/checkout@v2
-    
+
       # Update docker-compose
       - name: Update docker-compose on the server
         uses: appleboy/scp-action@master
