+++ backend/common/requirements.txt
@@ -1,2 +1,3 @@
+alembic
 pydantic
 sqlalchemy
