+++ backend/common/steam_database/alembic/alembic.ini
@@ -2,7 +2,7 @@
 
 [alembic]
 # path to migration scripts
-script_location = alembic
+script_location = common/steam_database/alembic
 
 # template used to generate migration files
 # file_template = %%(rev)s_%%(slug)s
@@ -17,7 +17,7 @@ prepend_sys_path = .
 # installed by adding `alembic[tz]` to the pip requirements
 # string value is passed to dateutil.tz.gettz()
 # leave blank for localtime
-# timezone = 
+# timezone =
 
 # max length of characters to apply to the
 # "slug" field
