--- backend/database_migration/requirements.txt
@@ -1,1 +1,0 @@
-alembic
